+++ src/main/kotlin/com/ampnet/crowdfundingbackend/service/WalletService.kt
@@ -3,10 +3,9 @@ package com.ampnet.crowdfundingbackend.service
 import com.ampnet.crowdfundingbackend.persistence.model.Project
 import com.ampnet.crowdfundingbackend.persistence.model.User
 import com.ampnet.crowdfundingbackend.persistence.model.Wallet
-import java.math.BigDecimal
 
 interface WalletService {
-    fun getWalletBalance(wallet: Wallet): BigDecimal
+    fun getWalletBalance(wallet: Wallet): Long
     fun createUserWallet(user: User, address: String): Wallet
     fun createProjectWallet(project: Project, address: String): Wallet
 }
